browser/task-form: subcomponents
