+++ browser/src/components/task_creation_form/containers/FloatContainer.vue
@@ -1,0 +1,15 @@
+<template>
+  <CustomField :field="props.field" />
+</template>
+
+<script lang="ts" setup>
+import { defineProps } from 'vue'
+
+import { FormField } from '@/task_creation_form'
+import CustomField from '@/components/task_creation_form/simple/CustomField.vue'
+
+interface Props {
+    field: FormField
+}
+const props = defineProps<Props>()
+</script>
