+++ browser/src/router/router.ts
@@ -4,7 +4,7 @@ import TrainingBar from '@/components/progress_bars/TrainingBar.vue'
 import InformationBar from '@/components/progress_bars/InformationBar.vue'
 import ValidationBar from '@/components/progress_bars/ValidationBar.vue'
 import Home from '@/components/home/Home.vue'
-import NewTaskCreationForm from '@/components/pages/NewTaskCreationForm.vue'
+import TaskCreationForm from '@/components/task_creation_form/TaskCreationForm.vue'
 import TaskList from '@/components/pages/TaskList.vue'
 import NotFound from '@/components/pages/NotFound.vue'
 import Training from '@/components/training/Training.vue'
@@ -25,8 +25,8 @@ export const router = createRouter({
     },
     {
       path: '/create',
-      name: 'new-task-form',
-      component: NewTaskCreationForm
+      name: 'task-creation-form',
+      component: TaskCreationForm
     },
     {
       path: '/information',
